Include changes from #999
